Simulate from simple individual-based model

Simulate from simple individual-based model

sim_indiv(max_time = 365, a = 0.3, p = 0.9, mu = -log(p), u = 22,
  v = 10, g = 10, r = 1/20, b = 1, c = 1, Ih_init = 10, H = 1000,
  M = 500, max_infections = 5, migration_matrix = matrix(1),
  H_auto = FALSE)

Arguments

max_time
run simulation for this many days

a
human blood feeding rate. The proportion of mosquitoes that feed on humans each day.

p
mosquito probability of surviving one day.

mu
mosquito instantaneous death rate. mu = -log(p) unless otherwise specified.

u
intrinsic incubation period. The number of days from infection to becoming infectious in a human host.

v
extrinsic incubation period. The number of days from infection to becoming infectious in a mosquito.

g
lag time between human bloodstage infection and production of gametocytes.

r
daily recovery rate.

b
probability a human becomes infected after being bitten by an infected mosquito.

c
probability a mosquito becomes infected after biting an infected human.

Ih_init
initial number of infectious humans in each deme.

H
human population size in each deme.

M
mosquito population size (number of adult female mosquitoes) in each deme.

max_infections
maximum number of infections that an individual can hold simultaneously.

migration_matrix
matrix giving the probability of migrating from current deme (rows) to new deme (columns).

H_auto
if TRUE then human population sizes are chosen automatically based on the migration matrix. In this case the total number of humans is equal to sum(H).

Details

TODO

Examples

# TODO